requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-limiter==3.7.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
